Fixed API exposed by package
