+++ pyproject.toml
@@ -33,7 +33,6 @@ dependencies = [
     "multivolumefile",
     "opencv-python",
     "pillow",
-    "pip",
     "psutil",
     "pyarrow",
     "pyyaml",
@@ -44,7 +43,6 @@ dependencies = [
     "python-dotenv",
     "pyzstd",
     "requests",
-    "setuptools",
     "texttable",
     "typing-extensions",
     "wheel"
